requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
mysqlclient==2.0.1
Pillow==8.0.1
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
